# Faster ensure_string_array