Benchmarks are still using 0.29.x, but I think this also prevents a regression from upgrading to 3.X as noted in #55179 

```sh
| Change   | Before [f4f598fb] <main>   | After [b12de6c8] <faster-ensure-string>   |   Ratio | Benchmark (Parameter)                                                          |
|----------|----------------------------|-------------------------------------------|---------|--------------------------------------------------------------------------------|
| -        | 1.78±0.02ms                | 1.42±0.01ms                               |    0.8  | strings.Construction.time_construction('categorical_series', 'string[python]') |
| -        | 981±10μs                   | 694±6μs                                   |    0.71 | strings.Construction.time_construction('frame', 'string[python]')              |
| -        | 860±7μs                    | 436±4μs                                   |    0.51 | strings.Construction.time_construction('series', 'string[python]')             |

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.
```