@rhshadrach 2023-10-20
This patch may have induced a performance regression. If it was a necessary behavior change, this may have been expected and everything is okay.

Please check the links below. If any ASVs are parameterized, the combinations of parameters that a regression has been detected for appear as subbullets.

 - [libs.InferDtype.time_infer_dtype](https://asv-runner.github.io/asv-collection/pandas/#libs.InferDtype.time_infer_dtype)
   - [dtype='np-object'](https://asv-runner.github.io/asv-collection/pandas/#libs.InferDtype.time_infer_dtype?p-dtype=%27np-object%27)
 - [libs.InferDtype.time_infer_dtype_skipna](https://asv-runner.github.io/asv-collection/pandas/#libs.InferDtype.time_infer_dtype_skipna)
   - [dtype='np-object'](https://asv-runner.github.io/asv-collection/pandas/#libs.InferDtype.time_infer_dtype_skipna?p-dtype=%27np-object%27)

Subsequent benchmarks may have skipped some commits. The link below lists the commits that are between the two benchmark runs where the regression was identified.

https://github.com/pandas-dev/pandas/compare/28f274cb107ef0a252eb4cb954c1cfbe43968395...2a08b052305ed35150d1355bc537548bec697778